+++ service/common/src/main/java/org/apache/polaris/service/catalog/iceberg/DefaultCatalogPrefixParser.java
@@ -20,9 +20,10 @@ package org.apache.polaris.service.catalog.iceberg;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import org.apache.polaris.core.context.RealmContext;
+import org.apache.polaris.service.catalog.CatalogPrefixParser;
 
 @ApplicationScoped
-public class DefaultIcebergCatalogPrefixParser implements IcebergCatalogPrefixParser {
+public class DefaultCatalogPrefixParser implements CatalogPrefixParser {
   @Override
   public String prefixToCatalogName(RealmContext realm, String prefix) {
     return prefix;
