+++ service/common/src/testFixtures/java/org/apache/polaris/service/TestServices.java
@@ -46,7 +46,7 @@ import org.apache.polaris.service.admin.PolarisServiceImpl;
 import org.apache.polaris.service.admin.api.PolarisCatalogsApi;
 import org.apache.polaris.service.catalog.api.IcebergRestCatalogApi;
 import org.apache.polaris.service.catalog.api.IcebergRestCatalogApiService;
-import org.apache.polaris.service.catalog.iceberg.DefaultIcebergCatalogPrefixParser;
+import org.apache.polaris.service.catalog.iceberg.DefaultCatalogPrefixParser;
 import org.apache.polaris.service.catalog.iceberg.IcebergCatalogAdapter;
 import org.apache.polaris.service.catalog.io.FileIOFactory;
 import org.apache.polaris.service.catalog.io.MeasuredFileIOFactory;
@@ -181,7 +181,7 @@ public record TestServices(
               configurationStore,
               polarisDiagnostics,
               authorizer,
-              new DefaultIcebergCatalogPrefixParser());
+              new DefaultCatalogPrefixParser());
 
       IcebergRestCatalogApi restApi = new IcebergRestCatalogApi(service);
